requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.8.0=py37_0
alabaster=0.7.12=py_0
asn1crypto=1.2.0=py37_0
astor=0.8.0=py37_0
astroid=2.3.2=py37_0
attrs=19.2.0=py_0
babel=2.7.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
bzip2=1.0.8=he774522_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
cffi=1.13.1=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.0=py_0
click-plugins=1.1.1=py_0
cligj=0.5.0=py37_0
cloudpickle=1.2.2=py_0
colorama=0.4.1=py37_0
cryptography=2.8=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
curl=7.67.0=h2a8f88b_0
cycler=0.10.0=py37_0
decorator=4.4.0=py37_1
defusedxml=0.6.0=py_0
docutils=0.15.2=py37_0
entrypoints=0.3=py37_0
expat=2.2.5=he025d50_0
fiona=1.8.4=py37h22081e2_0
freetype=2.9.1=ha9979f8_1
freexl=1.0.5=hfa6e2cd_0
future=0.18.2=pypi_0
gast=0.3.2=py_0
gdal=2.3.3=py37hdf43c64_0
geopandas=0.6.1=py_0
geos=3.7.1=h33f27b4_0
grpcio=1.16.1=py37h351948d_1
gym=0.15.4=pypi_0
h5py=2.9.0=py37h5e291fa_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py37_0
imagesize=1.1.0=py_0
intel-openmp=2019.4=245
ipdb=0.12.3=py_0
ipykernel=5.1.2=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
joblib=0.13.2=py37_0
jpeg=9b=hb83a4c4_2
jsonschema=3.0.2=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.0=py37_0
kealib=1.4.7=h07cbb95_6
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-preprocessing=1.1.0=py_1
keyring=18.0.0=py37_0
kiwisolver=1.1.0=py37ha925a31_0
krb5=1.16.4=hc04afaa_0
lazy-object-proxy=1.4.2=py37he774522_0
libboost=1.67.0=hd9e427e_4
libcurl=7.67.0=h2a8f88b_0
libgdal=2.3.3=h10f50ba_0
libiconv=1.15=h1df5818_7
libkml=1.3.0=he5f2a48_4
libmklml=2019.0.5=0
libnetcdf=4.6.1=h411e497_2
libpng=1.6.37=h2a8f88b_0
libpq=11.2=h3235a2c_0
libprotobuf=3.9.2=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libspatialite=4.3.0a=hc36aec2_19
libssh2=1.8.2=h7a1dbc1_0
libtiff=4.1.0=h56a325e_0
libxml2=2.9.9=h464c3ec_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
nbconvert=5.6.0=py37_1
nbformat=4.4.0=py37_0
ninja=1.9.0=py37h74a9793_0
notebook=6.0.1=py37_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
numpydoc=0.9.1=py_0
opencv-python=4.1.2.30=pypi_0
openssl=1.1.1d=he774522_3
packaging=19.2=py_0
pandas=0.25.1=py37ha925a31_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pcre=8.43=ha925a31_0
pickleshare=0.7.5=py37_0
pip=19.2.3=py37_0
proj4=5.2.0=ha925a31_1
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.9.2=py37h33f27b4_0
psutil=5.6.3=py37he774522_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pyflakes=2.1.1=py37_0
pyglet=1.3.2=pypi_0
pygments=2.4.2=py_0
pylint=2.4.3=py37_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.2=py_0
pyproj=1.9.6=py37h6782396_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.15.4=py37he774522_0
pysocks=1.7.1=py37_0
python=3.7.4=h5263a28_0
python-dateutil=2.8.0=py37_0
pytorch=1.3.1=py3.7_cuda101_cudnn7_0
pytz=2019.3=py_0
pywin32=223=py37hfa6e2cd_1
pywinpty=0.5.5=py37_1000
pyyaml=5.1.2=py37he774522_0
pyzmq=18.1.0=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.6.0=py_0
qtconsole=4.5.5=py_0
qtpy=1.9.0=py_0
requests=2.22.0=py37_0
rope=0.14.0=py_0
rtree=0.9.3=py37h21ff451_0
scikit-learn=0.21.3=py37h6288b17_0
scipy=1.3.1=py37h29ff71c_0
send2trash=1.5.0=py37_0
setuptools=41.4.0=py37_0
shapely=1.6.4=py37h222a598_0
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
snowballstemmer=2.0.0=py_0
sphinx=2.2.1=py_0
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
spyder=3.3.6=py37_0
spyder-kernels=0.5.2=py37_0
sqlite=3.30.0=he774522_0
tensorboard=1.14.0=py37he3c9ec2_0
tensorflow=1.14.0=mkl_py37h7908ca0_0
tensorflow-base=1.14.0=mkl_py37ha978198_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=py37_1
terminado=0.8.2=py37_0
testpath=0.4.2=py37_0
tk=8.6.8=hfa6e2cd_0
torchcontrib=0.0.2=pypi_0
tornado=6.0.3=py37he774522_0
traitlets=4.3.3=py37_0
urllib3=1.24.2=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.11.2=py37he774522_0
xerces-c=3.2.2=ha925a31_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zeromq=4.3.1=h33f27b4_3
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
